refactor(services): split pipelinesService monolith into focused modules (COD-117, 2/2)
Ports the fork's structural refactor of pipelinesService onto develop with zero
behavior change — all 206 pre-existing service tests pass unmodified:

- prompts.ts: the three LLM system-prompt builders (generate/analyze/optimize)
  extracted as pure functions; byte-identical to the previous inline strings,
  now locked by snapshot tests (+6).
- promptText.ts: shared truncation helpers.
- runStructuredAgent: the "run agent → extract JSON → retry" harness shared by
  all structured agent calls, with reason codes (AGENT_FAILED/BAD_AGENT_OUTPUT);
  optimizeFromDistillation now degrades gracefully on malformed agent JSON
  instead of throwing through an unguarded JSON.parse.
- proposeActions.ts: develop's propose implementation moved verbatim into its
  own module (behavior unchanged); the propose-engine expansion arrives with
  the Propose Engine issue on this isolated surface.
- createPipelinesService.ts shrinks from 1,267 to ~500 lines.

Deviation from the fork: the runtimeType? parameter on analyzeIntent and
generateStructure is kept (the fork dropped it) — develop's
pipelineAgentSessionsService passes it; dropping it fails repo typecheck.

Also duplicates PR #122's one-line db-schema test guard so repo-wide
check-types runs on this branch too (merges cleanly with either order).

Verification: services 212/212 tests, turbo check-types 19/19.
