+++ packages/services/src/pipelinesService/index.ts
@@ -1,1 +1,2 @@
 export * from "./createPipelinesService";
+export * from "./proposeActions";
